moved animation importer to compiler

+++ Borealis/inc/Graphics/Animation/Animation.hpp
@@ -28,7 +28,7 @@ namespace Borealis
 		void Save();
 
 		float mDuration;
-		int mTicksPerSecond;
+		float mTicksPerSecond;
 		std::vector<Bone> mBones;
 		AssimpNodeData mRootNode;
 		std::map<std::string, BoneData> mBoneDataMap;
